projet_segmentation/requirements.txt
streamlit
pillow
numpy
scikit-learn
matplotlib